Adding the code for "Investment Performance Web API" exercise
This is the project developed for the coding exercise.

The distribution package contains everything to start the
Web API engine in Linux or in Windows environment.

Assumptions
===========
I didn't add any database engine, instead I used `json` files
to store locally (in the `resources` folder) the data.
In the code, I provided SQL statement examples that can be used
if a database connection is required.

The code is structured in layers so moving from one source of
data to another one should be easy.
I provided a set of unit-tests to cover some coding aspects.
In a production ready code the number of unittests should be
significantly greater and should cover all the service and
manager layers public and protected classes.
I also covered some code with try and catch blocks and I
thrown some exceptions.

I know that making sure the code is stable is important, and
I assumed that this coding exercise is meant to provide a
general idea about coding techniques and project design.

Please consider that I am aware that many more coverage is needed
in terms of try/catch blocks, specialized exceptions and tests.

The project is written in Scala with gradle as builder and automation tool.

Tibor Dumitriu
tdumitriu@gmail.com

August 04, 2020